Clean up hint boxes and move View docs
The View docs file should be more consistent with Collection and Graph.

The hint boxes previously implied the arangojs API would be different depending on which ArangoDB version arangojs targets, which is not the case for the methdos in question. Since the views API was introduced in 3.4 it also doesn't make sense to repeat the hint box for every method on the view instance.